added a get request to get all user resumes
