+++ src/app/api/resumes/route.ts
@@ -1,0 +1,23 @@
+import { prisma } from "@/lib/prisma";
+
+export async function GET(request: Request) {
+    const { searchParams } = new URL(request.url);
+    const userId = searchParams.get("userId");
+    if (!userId) {
+        return new Response(JSON.stringify({ error: "No userId provided" }), {
+            status: 400,
+        });
+    }
+    try {
+        const resumes = await prisma.resume.findMany({
+            where: {
+                userId: userId,
+            },
+        });
+        return new Response(JSON.stringify(resumes), { status: 200 });
+    } catch (error: unknown) {
+        return new Response(JSON.stringify({ error: "An error occurred" }), {
+            status: 500,
+        });
+    }
+}
