+++ src/features/dashboard/components/resumes/Resumes.module.css
@@ -1,3 +1,8 @@
 .container {
     width: 100%;
 }
+
+.resumesContainer {
+    width: 100%;
+    background-color: red;
+}
