+++ src/features/dashboard/components/resumes/Resumes.tsx
@@ -1,30 +1,49 @@
 import styles from "./Resumes.module.css";
-import Card from "@/features/dashboard/components/cards/dashboard-cards/Card";
+import { DashboardCard } from "@/features/dashboard";
 import Link from "next/link";
 import { UpdateUrl } from "@/lib/updateUrl";
+import { getAllUserResumes } from "@/features/editor";
+import Basic from "@/components/resume-templates/basic/Basic";
 
 type ResumesProps = {
     currentUser: any;
     searchParams?: { [key: string]: string | string[] | undefined };
 };
 
-const Resumes = ({ currentUser, searchParams }: ResumesProps) => {
+const Resumes = async ({ currentUser, searchParams }: ResumesProps) => {
+    const resumes = await getAllUserResumes(currentUser.id);
+    console.log(resumes);
+
+    const addResumeButton = (
+        <Link
+            href={UpdateUrl(
+                searchParams ? searchParams : {},
+                [
+                    { key: "menu", value: "resumes" },
+                    { key: "createResume", value: "true" },
+                ],
+                "/dashboard"
+            )}
+        >
+            Add Resume
+        </Link>
+    );
     return (
         <main className={styles.container}>
-            <Card key="resumes-card" title="Resumes">
-                <Link
-                    href={UpdateUrl(
-                        searchParams ? searchParams : {},
-                        [
-                            { key: "menu", value: "resumes" },
-                            { key: "createResume", value: "true" },
-                        ],
-                        "/dashboard"
-                    )}
-                >
-                    Add Resume
-                </Link>
-            </Card>
+            <DashboardCard key="resumes-card" title="Resumes">
+                <section className={styles.resumesContainer}>
+                    {resumes.map((resume: any) => {
+                        return (
+                            <section
+                                className={styles.resumeCard}
+                                key={resume.id}
+                            >
+                                {resume.information.sectionOrder}
+                            </section>
+                        );
+                    })}
+                </section>
+            </DashboardCard>
         </main>
     );
 };
