+++ src/features/editor/components/page-util-bar/PageUtilBar.module.css
@@ -1,7 +1,7 @@
 .container {
     width: 70%;
-    min-width: 309px;
-    max-width: 850px;
+    min-width: 300px;
+    max-width: 800px;
     margin-inline: auto;
     margin-block-end: var(--padding-block-small);
     display: flex;
