+++ src/features/editor/index.ts
@@ -5,8 +5,12 @@ export { default as PageUtilBar } from "./components/page-util-bar/PageUtilBar";
 export { default as PageCounter } from "./components/page-util-bar/page-counter/PageCounter";
 export { default as SideMenu } from "./components/side-menu/SideMenu";
 export { default as TitleBar } from "./components/title-bar/TitleBar";
-export { updateDocumentArray } from "./lib/document";
 export { SectionConfig } from "./lib/sectionConfig";
-export { getResume, initializeNewResume } from "./lib/resume";
+export {
+    updateDocumentArray,
+    getResume,
+    initializeNewResume,
+    getAllUserResumes,
+} from "./lib/document";
 export { createResume as createResumeAction } from "./actions/document";
 export { updateResume as updateResumeAction } from "./actions/document";
